requirements.txt
discord
yt_dlp
pytz
python-dotenv